requirements.txt
fastapi
requests
python-dotenv
uvicorn[standard]
psycopg[binary]>=3.1
